Standardize chainid maps for oracles
Implements standardized chainid maps for oracles. This also lets
oracles opt out of mapping in case they conform to a chainid
standardization. This applies to Polymer for EVM chains. The
advantage of non-mapped oracles is that they are ownerless making the
system entirely permissionless and providing a base implementation
that is entirely configless: **One deterministic deployment**